회원가입 및 로그인 기능 구현

+++ src/main/kotlin/com/tik/server/entity/TechStack.kt
@@ -10,7 +10,7 @@ class TechStack(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    val id: Long = 0
+    val id: Int = 0
 
     @Column(name = "name")
     var name: String = name
